Refactor and move App and AppLayout inside the "components" directory
