+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/App.js
@@ -17,18 +17,18 @@
  */
 
 import AppLayout from "./AppLayout";
-import Cells from "./components/cells";
-import {ColorProvider} from "./components/common/color";
-import ErrorBoundary from "./components/common/error/ErrorBoundary";
-import NotFound from "./components/common/error/NotFound";
-import Overview from "./components/overview";
+import Cells from "./cells";
+import {ColorProvider} from "./common/color";
+import ErrorBoundary from "./common/error/ErrorBoundary";
+import NotFound from "./common/error/NotFound";
+import Overview from "./overview";
 import React from "react";
-import SignIn from "./components/SignIn";
-import SystemMetrics from "./components/systemMetrics";
-import Tracing from "./components/tracing";
+import SignIn from "./SignIn";
+import SystemMetrics from "./systemMetrics";
+import Tracing from "./tracing";
 import {BrowserRouter, Route, Switch} from "react-router-dom";
 import {MuiThemeProvider, createMuiTheme} from "@material-ui/core/styles";
-import withGlobalState, {StateHolder, StateProvider} from "./components/common/state";
+import withGlobalState, {StateHolder, StateProvider} from "./common/state";
 import * as PropTypes from "prop-types";
 
 class StatelessProtectedPortal extends React.Component {
