+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/AppLayout.js
@@ -19,15 +19,15 @@
 
 import AccountCircle from "@material-ui/icons/AccountCircle";
 import AppBar from "@material-ui/core/AppBar";
-import AuthUtils from "./utils/api/authUtils";
-import CellsIcon from "./icons/CellsIcon";
+import AuthUtils from "../utils/api/authUtils";
+import CellsIcon from "../icons/CellsIcon";
 import CheckCircle from "@material-ui/icons/CheckCircle";
 import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import CircularProgress from "@material-ui/core/CircularProgress";
 import CloseIcon from "@material-ui/icons/Close";
 import Collapse from "@material-ui/core/Collapse";
-import ColorGenerator from "./components/common/color/colorGenerator";
+import ColorGenerator from "./common/color/colorGenerator";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import Divider from "@material-ui/core/Divider";
 import Drawer from "@material-ui/core/Drawer";
@@ -46,16 +46,16 @@ import ListItemText from "@material-ui/core/ListItemText";
 import Menu from "@material-ui/core/Menu";
 import MenuIcon from "@material-ui/icons/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
-import MetricsIcon from "./icons/MetricsIcon";
-import NodeIcon from "./icons/NodeIcon";
-import NotificationUtils from "./utils/common/notificationUtils";
-import OverviewIcon from "./icons/OverviewIcon";
+import MetricsIcon from "../icons/MetricsIcon";
+import NodeIcon from "../icons/NodeIcon";
+import NotificationUtils from "../utils/common/notificationUtils";
+import OverviewIcon from "../icons/OverviewIcon";
 import Paper from "@material-ui/core/Paper/Paper";
-import PodIcon from "./icons/PodIcon";
+import PodIcon from "../icons/PodIcon";
 import Popover from "@material-ui/core/Popover";
 import React from "react";
 import Settings from "@material-ui/icons/SettingsOutlined";
-import ShareIcon from "./icons/ShareIcon";
+import ShareIcon from "../icons/ShareIcon";
 import Snackbar from "@material-ui/core/Snackbar/Snackbar";
 import Timeline from "@material-ui/icons/Timeline";
 import Toolbar from "@material-ui/core/Toolbar";
@@ -63,10 +63,10 @@ import Tooltip from "@material-ui/core/Tooltip";
 import Typography from "@material-ui/core/Typography";
 import Warning from "@material-ui/icons/Warning";
 import classNames from "classnames";
-import withColor from "./components/common/color";
+import withColor from "./common/color";
 import {withRouter} from "react-router-dom";
 import {withStyles} from "@material-ui/core/styles";
-import withGlobalState, {StateHolder} from "./components/common/state";
+import withGlobalState, {StateHolder} from "./common/state";
 import * as PropTypes from "prop-types";
 
 const drawerWidth = 240;
