+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/index.js
@@ -17,7 +17,7 @@
  */
 
 import "./index.css";
-import App from "./App";
+import App from "./components/App";
 import React from "react";
 import ReactDOM from "react-dom";
 import * as serviceWorker from "./serviceWorker";
